+++ NEWS.md
@@ -1,3 +1,8 @@
+# Versão 0.2.1 (2019-03-09)
+* Renomeie o Id do downloadButton de 'download' para 'outDown'
+* Erro do download resolvido (alterei 'file = filename' para  'file')
+* Alterei 'filename <- function' para 'filename = function'
+
 # Versão 0.2.0 (2019-03-08)
 * Inclui documento `NEWS.md` para registro das versões da aplicação.
 * Altera o nome do documento de código principal para `app.R`.